+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Move `Qless\Listener` to the `Qless\Events\Subscriber`
 - More code quality improvements to follow the SRP. Thus the code base for almost all classes has been changed
 - Move all the exceptions to the common namespace and implement the same `Qless\Exceptions\ExceptionInterface`
+- Changed `Qless\Queue::put` signature from the `put($className, $jid, $data, ...)` to the `put(string className, array $data, ?string $jid = null, ...)`
   
 ### Removed
 - Fully refactor the `Qless\Client` class and removed no longer used code